refactor(routes): update all mod-view routes to mod-portal
