+++ src/app/(protected)/mod-portal/layout.jsx
@@ -1,8 +1,9 @@
 "use client";
-import React from "react";
-import { ModNavigation } from "@/components/ModView/ModNavigation/ModNavigation";
 
-export default function ModViewLayout({ children }) {
+import { ModNavigation } from "@/components/ModPortal/ModNavigation/ModNavigation";
+import React from "react";
+
+export default function ModPortalLayout({ children }) {
   return (
     <div className="grid grid-cols-1 lg:grid-cols-[25ch_1fr] max-w-screen-2xl mx-auto py-4">
       <div className="scrollbar">
