+++ src/app/(protected)/mod-portal/notes/page.js
@@ -1,7 +1,7 @@
 "use client";
 
-import { ModViewDataGrid } from "@/components/ModView/ModViewDataGrid/ModViewDataGrid";
-import { makeData } from "@/components/ModView/makeData";
+import { ModPortalDataGrid } from "@/components/ModPortal/ModPortalDataGrid/ModPortalDataGrid";
+import { makeData } from "@/components/ModPortal/makeData";
 
 const testData = makeData(100, ["userId", "issuedBy", "lastEdited", "notes"]);
 
@@ -13,12 +13,12 @@ const columns = [
 ];
 
 /**
- * Mod View page for Notes
+ * Mod Portal page for Notes
  * */
-export default function ModViewNotes(props) {
+export default function ModPortalNotes(props) {
   return (
     <div>
-      <ModViewDataGrid columns={columns} rows={testData} />
+      <ModPortalDataGrid columns={columns} rows={testData} />
     </div>
   );
 }
