+++ src/app/(protected)/mod-portal/page.js
@@ -1,0 +1,8 @@
+"use client";
+
+/**
+ * Mod Portal root page
+ * */
+export default function ModPortal(props) {
+  return <div></div>;
+}
