+++ src/app/(protected)/mod-portal/settings/page.js
@@ -1,0 +1,8 @@
+"use client";
+
+/**
+ * Mod Portal page for Settings
+ * */
+export default function ModPortalSettings(props) {
+  return <div></div>;
+}
