--- src/app/mod-view/page.js
@@ -1,8 +1,0 @@
-"use client";
-
-/**
- * Mod View root page
- * */
-export default function ModView(props) {
-  return <div></div>;
-}
