--- src/app/mod-view/settings/page.js
@@ -1,8 +1,0 @@
-"use client";
-
-/**
- * Mod View page for Settings
- * */
-export default function ModViewSettings(props) {
-  return <div></div>;
-}
